+++ Modules/Intent.Modules.AspNetCore.Controllers.Dispatch.MediatR/FactoryExtensions/MediatRControllerInstaller.cs
@@ -64,7 +64,7 @@ namespace Intent.Modules.AspNetCore.Controllers.Dispatch.MediatR.FactoryExtensio
                         {
                             method.AddStatements(GetValidations(model));
                             method.AddStatement(GetDispatchViaMediatorStatement(template, model), s => s.SeparatedFromPrevious());
-                            method.AddStatement(GetReturnStatement(template, model));
+                            method.AddStatement(template.GetReturnStatement(model));
                         }
                     }
                 });
@@ -99,67 +99,6 @@ namespace Intent.Modules.AspNetCore.Controllers.Dispatch.MediatR.FactoryExtensio
             return operationModel.ReturnType != null
                 ? $"var result = await _mediator.Send({payload}, cancellationToken);"
                 : $"await _mediator.Send({payload}, cancellationToken);";
-        }
-
-        private static CSharpStatement GetReturnStatement(ControllerTemplate template, IControllerOperationModel operationModel)
-        {
-            var hasReturnType = operationModel.ReturnType != null;
-
-            var resultExpression = default(string);
-            if (hasReturnType)
-            {
-                resultExpression = template.ShouldBeJsonResponseWrapped(operationModel)
-                    ? $"new {template.GetJsonResponseName()}<{template.GetTypeName(operationModel.ReturnType)}>(result)"
-                    : "result";
-            }
-
-            switch (operationModel.Verb)
-            {
-                case HttpVerb.Get:
-                case HttpVerb.Patch:
-                case HttpVerb.Put:
-                    return hasReturnType ? $"return Ok({resultExpression});" : "return NoContent();";
-                case HttpVerb.Delete:
-                    return hasReturnType ? $"return Ok({resultExpression});" : "return Ok();";
-                case HttpVerb.Post:
-                    switch (operationModel.Parameters.Count)
-                    {
-                        case 1:
-                            // Aggregate
-                        {
-                            var getByIdOperation = template.Model.Operations.FirstOrDefault(x => x.Verb == HttpVerb.Get &&
-                                x.ReturnType is { IsCollection: false } &&
-                                x.Parameters.Count == 1 &&
-                                string.Equals(x.Parameters[0].Name, "id", StringComparison.OrdinalIgnoreCase));
-                            if (getByIdOperation != null && operationModel.ReturnType?.Element.Name is "guid" or "long" or "int" or "string")
-                            {
-                                return $"return CreatedAtAction(nameof({getByIdOperation.Name}), new {{ id = result }}, {resultExpression});";
-                            }
-                        }
-                            break;
-                        case 2:
-                            // Owned composite
-                        {
-                            var getByIdOperation = template.Model.Operations.FirstOrDefault(x => x.Verb == HttpVerb.Get &&
-                                x.ReturnType is { IsCollection: false } &&
-                                x.Parameters.Count == 2 &&
-                                string.Equals(x.Parameters[0].Name, operationModel.Parameters[0].Name, StringComparison.OrdinalIgnoreCase) &&
-                                string.Equals(x.Parameters[1].Name, "id", StringComparison.OrdinalIgnoreCase));
-                            if (getByIdOperation != null && operationModel.ReturnType?.Element.Name is "guid" or "long" or "int" or "string")
-                            {
-                                var aggregateIdParameter = getByIdOperation.Parameters[0].Name.ToCamelCase();
-                                return $"return CreatedAtAction(nameof({getByIdOperation.Name}), new {{ {aggregateIdParameter} = {aggregateIdParameter}, id = result }}, {resultExpression});";
-                            }
-                        }
-                            break;
-                        default:
-                            break;
-                    }
-
-                    return hasReturnType ? $"return Created(string.Empty, {resultExpression});" : "return Created(string.Empty, null);";
-                default:
-                    throw new ArgumentOutOfRangeException();
-            }
         }
 
         private static IControllerParameterModel GetPayloadParameter(IControllerOperationModel operation)